requirements.txt
Django~=5.1.5
pyotp~=2.9.0
python-dotenv~=1.0.1
django-allauth~=65.4.1
django-import-export~=4.3.5
django-nested-admin~=4.1.1